+++ packages/element-scheduler/src/Worker.ts
@@ -9,6 +9,7 @@ import {
 	ParentMessageError,
 	ParentMessageOk,
 	MessageConst,
+	OnReport,
 } from './types'
 import { join } from 'path'
 import { Worker as WorkerThread } from 'worker_threads'
@@ -16,6 +17,8 @@ import { PassThrough } from 'stream'
 import mergeStream from 'merge-stream'
 import { Page } from 'playwright'
 import { WorkerConnection } from './WorkerConnection'
+import debugFactory from 'debug'
+const debug = debugFactory('element:worker')
 
 export type WorkerOptions = {
 	page?: Promise<Page>
@@ -34,6 +37,7 @@ export class Worker implements WorkerInterface {
 
 	private retries = 0
 	private onProcessEnd: OnEnd
+	private onProcessReport: OnReport
 
 	private exitPromise: Promise<void>
 	private resolveExitPromise!: () => void
@@ -133,6 +137,8 @@ export class Worker implements WorkerInterface {
 					this.onProcessEnd(null, this, data[0] as number)
 				} else if (result === MessageConst.LOADED) {
 					this.resolveLoadPromise()
+				} else if (result === MessageConst.REPORT) {
+					this.onProcessReport(this, data as string[])
 				}
 				break
 			}
@@ -161,7 +167,7 @@ export class Worker implements WorkerInterface {
 	}
 
 	private onExit = (exitCode: number) => {
-		console.log(`User ${this.workerName} exit: ${exitCode}`)
+		debug(`exit code: ${exitCode}`)
 		this.resolveExitPromise()
 	}
 
@@ -178,9 +184,15 @@ export class Worker implements WorkerInterface {
 		this.worker.terminate()
 	}
 
-	send(request: ChildMessage, onProcessStart: OnStart, onProcessEnd: OnEnd): void {
+	send(
+		request: ChildMessage,
+		onProcessStart: OnStart,
+		onProcessEnd: OnEnd,
+		onProcessReport: OnReport,
+	): void {
 		onProcessStart(this)
 		this.onProcessEnd = (...args) => onProcessEnd(...args)
+		if (onProcessReport) this.onProcessReport = (...args) => onProcessReport(...args)
 
 		this.retries = 0