+++ packages/element-scheduler/src/cli.ts
@@ -1,5 +1,6 @@
 import { Scheduler } from './Scheduler'
 import { mustCompileFile, ElementOptions, EvaluatedScript } from '@flood/element-core'
+import Spinnies from 'spinnies'
 
 export async function runCommandLine(opts: ElementOptions): Promise<void> {
 	const { testScript, runEnv, testSettingOverrides, headless, browser } = opts
@@ -11,22 +12,13 @@ export async function runCommandLine(opts: ElementOptions): Promise<void> {
 		...testSettingOverrides,
 		headless: headless,
 		browser: browser || evaluateScript.settings.browser,
+		verbose: opts.verbose,
 	})
 
-	const installSignalHandlers = true
+	const spinnies = new Spinnies()
+	runner.setSpinnies(spinnies)
 
-	if (installSignalHandlers) {
-		process.on('SIGINT', async () => {
-			await runner.stop()
-		})
-
-		process.once('SIGUSR2', async () => {
-			await runner.stop()
-			process.kill(process.pid, 'SIGUSR2')
-		})
-	}
-
-	console.debug(`Loading test script: ${testScript}`)
+	spinnies.add('initializing', { text: 'Initializing users ...' })
 
 	await runner.run(testScript)
 	await runner.stop()